azure_conn/pyproject.toml
[project]
name = "azure-conn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.20.0",
    "azure-storage-blob>=12.24.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.1",
]
